Updated calculations
I realized that a $3 Price per client isn't the correct measurement. I
was including users that didn't contribute, when in fact all that
matters is those who do. So Price is now at $15

Regardless, the plan of not going down the route of Kickstarter and
nonprofit setup, and instead building the prototype on free time, is
still in action.