Bad fix for ',",<, and > in string bug
Users could type a ' (for example) in the body in newargument and when
it shows in viewargument everything would be messed up. I couldn't
figure out a way to add a \ in front of those characters so I just wrote
code that removes those characters. This is a temporary fix.